refactor(discord): move runtime startup doctor helpers
